+++ data/serializer/src/main/java/tech/pegasys/teku/api/response/v1/beacon/PostDataFailureResponse.java
@@ -17,22 +17,42 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
+import java.util.Objects;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
-public class PostSyncCommitteeFailureResponse {
+public class PostDataFailureResponse {
   @Schema(type = "string", format = "uint64")
   public final UInt64 code;
 
   public final String message;
-  public final List<PostSyncCommitteeFailure> failures;
+  public final List<PostDataFailure> failures;
 
   @JsonCreator
-  public PostSyncCommitteeFailureResponse(
+  public PostDataFailureResponse(
       @JsonProperty("code") final int code,
       @JsonProperty("message") final String message,
-      @JsonProperty("failures") final List<PostSyncCommitteeFailure> failures) {
+      @JsonProperty("failures") final List<PostDataFailure> failures) {
     this.code = UInt64.valueOf(code);
     this.message = message;
     this.failures = failures;
+  }
+
+  @Override
+  public boolean equals(final Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    final PostDataFailureResponse that = (PostDataFailureResponse) o;
+    return Objects.equals(code, that.code)
+        && Objects.equals(message, that.message)
+        && Objects.equals(failures, that.failures);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(code, message, failures);
   }
 }