+++ validator/remote/src/test/java/tech/pegasys/teku/validator/remote/apiclient/OkHttpValidatorRestApiClientTest.java
@@ -43,8 +43,8 @@ import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.api.request.v1.validator.BeaconCommitteeSubscriptionRequest;
 import tech.pegasys.teku.api.response.v1.beacon.GetGenesisResponse;
 import tech.pegasys.teku.api.response.v1.beacon.GetStateValidatorsResponse;
-import tech.pegasys.teku.api.response.v1.beacon.PostSyncCommitteeFailure;
-import tech.pegasys.teku.api.response.v1.beacon.PostSyncCommitteeFailureResponse;
+import tech.pegasys.teku.api.response.v1.beacon.PostDataFailure;
+import tech.pegasys.teku.api.response.v1.beacon.PostDataFailureResponse;
 import tech.pegasys.teku.api.response.v1.beacon.ValidatorResponse;
 import tech.pegasys.teku.api.response.v1.validator.GetAggregatedAttestationResponse;
 import tech.pegasys.teku.api.response.v1.validator.GetAttestationDataResponse;
@@ -388,7 +388,7 @@ class OkHttpValidatorRestApiClientTest {
 
     mockWebServer.enqueue(new MockResponse().setResponseCode(SC_OK));
 
-    apiClient.sendSignedAttestation(attestation);
+    apiClient.sendSignedAttestations(List.of(attestation));
 
     RecordedRequest request = mockWebServer.takeRequest();
 
@@ -403,10 +403,17 @@ class OkHttpValidatorRestApiClientTest {
   public void sendSignedAttestation_WhenBadParameters_ThrowsIllegalArgumentException() {
     final Attestation attestation = schemaObjects.attestation();
 
-    mockWebServer.enqueue(new MockResponse().setResponseCode(SC_BAD_REQUEST));
+    final PostDataFailureResponse response =
+        new PostDataFailureResponse(
+            SC_BAD_REQUEST, "Computer said no", List.of(new PostDataFailure(UInt64.ZERO, "Bad")));
+    mockWebServer.enqueue(
+        new MockResponse().setResponseCode(SC_BAD_REQUEST).setBody(asJson(response)));
 
-    assertThatThrownBy(() -> apiClient.sendSignedAttestation(attestation))
-        .isInstanceOf(IllegalArgumentException.class);
+    assertThat(apiClient.sendSignedAttestations(List.of(attestation)))
+        .isPresent()
+        .get()
+        .usingRecursiveComparison()
+        .isEqualTo(response);
   }
 
   @Test
@@ -415,7 +422,7 @@ class OkHttpValidatorRestApiClientTest {
 
     mockWebServer.enqueue(new MockResponse().setResponseCode(SC_INTERNAL_SERVER_ERROR));
 
-    assertThatThrownBy(() -> apiClient.sendSignedAttestation(attestation))
+    assertThatThrownBy(() -> apiClient.sendSignedAttestations(List.of(attestation)))
         .isInstanceOf(RuntimeException.class)
         .hasMessageContaining("Unexpected response from Beacon Node API");
   }
@@ -687,11 +694,9 @@ class OkHttpValidatorRestApiClientTest {
 
   @Test
   void sendSyncCommitteeMessages_shouldReturnErrorsFromBadResponse() {
-    final PostSyncCommitteeFailureResponse response =
-        new PostSyncCommitteeFailureResponse(
-            SC_BAD_REQUEST,
-            "Computer said no",
-            List.of(new PostSyncCommitteeFailure(UInt64.ZERO, "Bad")));
+    final PostDataFailureResponse response =
+        new PostDataFailureResponse(
+            SC_BAD_REQUEST, "Computer said no", List.of(new PostDataFailure(UInt64.ZERO, "Bad")));
     mockWebServer.enqueue(
         new MockResponse().setResponseCode(SC_BAD_REQUEST).setBody(asJson(response)));